# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: PEPA — Platform Engineering & Pipeline Automator API
  description: |
    PEPA — Platform Engineering & Pipeline Automator API for managing services, workflows, scorecards,
    plugins, RBAC, and platform operations.
  version: "0.1.0"
  contact:
    name: PEPA Team
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8088/api/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /workflows:
    get:
      tags: [Workflows]
      summary: List workflows
      operationId: listWorkflows
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  workflows:
                    type: array
                    items:
                      $ref: "#/components/schemas/Workflow"
                  total: { type: integer }
    post:
      tags: [Workflows]
      summary: Create workflow
      operationId: createWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Workflow"
      responses:
        "201":
          description: Created
  /workflows/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Workflows]
      summary: Get workflow
      operationId: getWorkflow
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
    put:
      tags: [Workflows]
      summary: Update workflow
      operationId: updateWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Workflow"
      responses:
        "200":
          description: Updated
    delete:
      tags: [Workflows]
      summary: Delete workflow
      operationId: deleteWorkflow
      responses:
        "200":
          description: Deleted
  /workflows/{id}/execute:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Workflows]
      summary: Execute workflow
      operationId: executeWorkflow
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                parameters: { type: object }
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowExecution"
  /workflows/{id}/executions:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Workflows]
      summary: List workflow executions
      operationId: listExecutions
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  executions:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkflowExecution"

  # ── Plugins ─────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Workflow:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        tenant_id: { type: string, format: uuid }
        spec: { type: object }
        version: { type: integer }
        is_enabled: { type: boolean }
        created_at: { type: string, format: date-time }
    WorkflowExecution:
      type: object
      properties:
        id: { type: string, format: uuid }
        workflow_id: { type: string, format: uuid }
        status: { type: string }
        trigger_type: { type: string }
        triggered_by: { type: string, format: uuid }
        started_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
        duration_ms: { type: integer }
...